+++ core/src/main/java/com/zelaya/gulag/Main.java
@@ -1,32 +1,20 @@
 package com.zelaya.gulag;
 
-import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.Game;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.utils.ScreenUtils;
+import com.zelaya.gulag.pantallas.PantallaMenu;
+import com.zelaya.gulag.utiles.Render;
 
-/** {@link com.badlogic.gdx.ApplicationListener} implementation shared by all platforms. */
-public class Main extends ApplicationAdapter {
-    private SpriteBatch batch;
-    private Texture image;
+public class Main extends Game {
 
     @Override
     public void create() {
-        batch = new SpriteBatch();
-        image = new Texture("libgdx.png");
-    }
-
-    @Override
-    public void render() {
-        ScreenUtils.clear(0.15f, 0.15f, 0.2f, 1f);
-        batch.begin();
-        batch.draw(image, 140, 210);
-        batch.end();
+        Render.batch = new SpriteBatch();
+        setScreen(new PantallaMenu(this));
     }
 
     @Override
     public void dispose() {
-        batch.dispose();
-        image.dispose();
+        Render.batch.dispose();
     }
 }