+++ core/src/main/java/com/zelaya/gulag/utiles/Config.java
@@ -1,0 +1,6 @@
+package com.zelaya.gulag.utiles;
+
+public class Config {
+    public static final int ANCHO = 1280;
+    public static final int ALTO = 720;
+}