+++ core/src/main/java/com/zelaya/gulag/utiles/Recursos.java
@@ -1,0 +1,8 @@
+package com.zelaya.gulag.utiles;
+
+public class Recursos {
+
+    public static final String FONDOMENU = "fondos/fondo_menu.jpg";
+    public static final String FUENTEMENU = "fuentes/gulag.otf";
+
+}